# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProjectCRUD",
        "description": "API documentation for the ProjectCRUD application.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/": {
            "get": {
                "tags": [
                    "Welcome"
                ],
                "summary": "Display a listing of cars",
                "description": "Display a listing of the resource.",
                "operationId": "indexWelcome",
                "parameters": [
                    {
                        "name": "filter[day_repayment_min]",
                        "in": "query",
                        "description": "Minimum day repayment",
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "filter[day_repayment_max]",
                        "in": "query",
                        "description": "Maximum day repayment",
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "defaultImage": {
                                            "type": "string"
                                        },
                                        "isGuest": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/borrow/{car}": {
            "get": {
                "tags": [
                    "Welcome"
                ],
                "summary": "Borrow a car",
                "description": "Borrow a car.",
                "operationId": "borrowCar",
                "parameters": [
                    {
                        "name": "car",
                        "in": "path",
                        "description": "ID of the car to borrow",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "carmake": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "enum": [
                                                    "Audi",
                                                    "BMW"
                                                ]
                                            }
                                        },
                                        "carbodyshape": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "enum": [
                                                    "Sedan",
                                                    "SUV"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Welcome"
                ],
                "summary": "Update the specified resource in storage",
                "description": "Update the specified resource in storage.",
                "operationId": "updateBorrow",
                "parameters": [
                    {
                        "name": "car",
                        "in": "path",
                        "description": "ID of the car to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpsertBorrowRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Successful operation - Redirect to welcome",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UpsertBorrowRequest": {
                "title": "UpsertBorrowRequest",
                "description": "Request data for creating/updating a borrow.",
                "required": [
                    "rental_date",
                    "return_date"
                ],
                "properties": {
                    "make": {
                        "description": "The make of the car being borrowed.",
                        "type": "string",
                        "example": "Toyota"
                    },
                    "model": {
                        "description": "The model of the car being borrowed.",
                        "type": "string",
                        "example": "Camry"
                    },
                    "body_shape": {
                        "description": "The body shape of the car being borrowed.",
                        "type": "string",
                        "example": "Sedan"
                    },
                    "rental_date": {
                        "description": "The rental date for the car.",
                        "type": "string",
                        "format": "date",
                        "example": "2023-07-19"
                    },
                    "return_date": {
                        "description": "The return date for the car.",
                        "type": "string",
                        "format": "date",
                        "example": "2023-07-25"
                    }
                },
                "type": "object"
            },
...
        }
    }
}